separate base email builder
